pyproject.toml
[tool.poetry]
name = "llm-knowledge-agent"
version = "0.1.0"
description = ""
authors = ["kylestratis <kyle.stratis@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.1"
cohere = "^4.3.1"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start_agent = "llm_knowledge_agent"  # TODO: Set up entrypoint once runner is fleshed out